Remove JoinProbe interface
